# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  isAdmin: boolean;
}

export interface DayEntry {
  day: string;
  netSales: number;
  worked: boolean;
}

export interface AdminWeekRecord {
  id: string;
  label: string;
  chatterName: string;
  weekStart: string;
  days: DayEntry[];
  createdAt: string;
  userId: string;
  /** @nullable */
  userEmail: string | null;
  /** @nullable */
  userFirstName: string | null;
  /** @nullable */
  userLastName: string | null;
}

export interface AdminWeekList {
  weeks: AdminWeekRecord[];
}

export interface AdminUserRecord {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  isAdmin: boolean;
  createdAt: string;
  weekCount: number;
}

export interface AdminUserList {
  users: AdminUserRecord[];
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface WeekRecord {
  id: string;
  label: string;
  chatterName: string;
  weekStart: string;
  days: DayEntry[];
  createdAt: string;
}

export interface WeekRecordList {
  weeks: WeekRecord[];
}

export interface CreateWeekRequest {
  label: string;
  chatterName: string;
  weekStart: string;
  days: DayEntry[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};