# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: weeks
    description: Weekly salary tracking records
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /weeks:
    get:
      tags: [weeks]
      operationId: getWeeks
      summary: Get all week records for the current user
      responses:
        '200':
          description: List of week records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekRecordList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [weeks]
      operationId: createWeek
      summary: Create a new week record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWeekRequest'
      responses:
        '201':
          description: Week record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /weeks/{id}:
    put:
      tags: [weeks]
      operationId: updateWeek
      summary: Update a week record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWeekRequest'
      responses:
        '200':
          description: Week record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Week not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [weeks]
      operationId: deleteWeek
      summary: Delete a week record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Week deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Week not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isAdmin]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        isAdmin:
          type: boolean
    AdminWeekRecord:
      type: object
      required: [id, label, chatterName, weekStart, days, createdAt, userId, userEmail, userFirstName, userLastName]
      properties:
        id:
          type: string
        label:
          type: string
        chatterName:
          type: string
        weekStart:
          type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/DayEntry'
        createdAt:
          type: string
        userId:
          type: string
        userEmail:
          type: [string, 'null']
        userFirstName:
          type: [string, 'null']
        userLastName:
          type: [string, 'null']
    AdminWeekList:
      type: object
      required: [weeks]
      properties:
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/AdminWeekRecord'
    AdminUserRecord:
      type: object
      required: [id, email, firstName, lastName, isAdmin, createdAt, weekCount]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        isAdmin:
          type: boolean
        createdAt:
          type: string
        weekCount:
          type: number
    AdminUserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserRecord'
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DayEntry:
      type: object
      required: [day, netSales, worked]
      properties:
        day:
          type: string
        netSales:
          type: number
        worked:
          type: boolean
    WeekRecord:
      type: object
      required: [id, label, chatterName, weekStart, days, createdAt]
      properties:
        id:
          type: string
        label:
          type: string
        chatterName:
          type: string
        weekStart:
          type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/DayEntry'
        createdAt:
          type: string
    WeekRecordList:
      type: object
      required: [weeks]
      properties:
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/WeekRecord'
    CreateWeekRequest:
      type: object
      required: [label, chatterName, weekStart, days]
      properties:
        label:
          type: string
        chatterName:
          type: string
        weekStart:
          type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/DayEntry'