# lib/api-zod/src/generated/types/adminWeekList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminWeekRecord } from "./adminWeekRecord";

export interface AdminWeekList {
  weeks: AdminWeekRecord[];
}