# lib/api-zod/src/generated/types/weekRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DayEntry } from "./dayEntry";

export interface WeekRecord {
  id: string;
  label: string;
  chatterName: string;
  weekStart: string;
  days: DayEntry[];
  createdAt: string;
}