# lib/api-zod/src/generated/types/weekRecordList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WeekRecord } from "./weekRecord";

export interface WeekRecordList {
  weeks: WeekRecord[];
}